+++ resources/views/components/events/event-type-statistic-event-list.blade.php
@@ -6,11 +6,11 @@
         @endphp
         <li>
             @if($hasEventEditPermission)
-                <x-button-link href="{{route('event.edit', $event->id)}}"
+                <a href="{{route('event.edit', $event->id)}}"
                                title="Edit this event"
-                               class="bg-gray-800 text-white hover:bg-gray-700 focus:bg-gray-700 active:bg-gray-900 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2">
+                               class="btn-primary">
                     <i class="fa-regular fa-pen-to-square"></i>
-                </x-button-link>
+                </a>
             @endif
             <span>{{$event->start->formatDateOnly(true)}} - {{$event->title}}</span>
         </li>